Moving math package from external libarary
